requirements.txt
langchain
langchain-google-genai
google-generativeai>=0.8.5
Pillow
streamlit